receptor_ai/app/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
bcrypt==4.1.2
beanie==1.26.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.115.0
fastapi-users==13.0.0
fastapi-users-db-beanie==3.0.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.10
lazy-model==0.2.0
makefun==1.15.4
motor==3.6.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pwdlib==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
pymongo==4.9.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.38.5
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
yarl==1.11.1
